style: add visitor file to store types which implement ast.Visitor interface
